move v1() to v1_compat
